Renamed misc.py to examples.py.
