+++ colorpy/examples.py
